+++ components/buttons/TrashButton.tsx
