+++ screens/CreateExpenseScreen.tsx
@@ -2,7 +2,7 @@ import { useMutation } from "@apollo/client";
 import { useState, useEffect } from "react";
 import { CreateExpenseDocument, CreateExpenseMutation } from "../components/generated";
 import { RootStackScreenProps } from "../types";
-import { ExpenseEditForm, FormValues } from "../components/ExpenseEditForm";
+import { ExpenseEditForm, FormValues } from "../components/forms/ExpenseEditForm";
 import { useAuth } from "../hooks/useAuth";
 import moment from "moment";
