+++ screens/ReportsScreen.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { StyleSheet, View, Text } from 'react-native';
 import { RootTabScreenProps } from "../types";
 import { useAuth } from "../hooks/useAuth";
-import { Screen } from "../components/Screen";
+import { Screen } from "../components/forms/Screen";
 import Styles from '../constants/Styles';
 import { VictoryChart, VictoryLegend, VictoryPie } from 'victory-native';
 import { GetCategoriesDocument, GetCategoriesQuery, GetMonthBreakdownQueryVariables, MonthType } from '../components/generated';
@@ -28,7 +28,7 @@ export default function ReportsScreen({ navigation }: RootTabScreenProps<'Report
     useRefresh(() => refetch({ passwordHash, month: month as MonthType, year }), [month, year]);
 
     return (
-        <Screen>
+        <View style={styles.screen}>
             <ScrollView>
                 <TopBar month={month} year={year} setMonth={setMonth} setYear={setYear} />
                 <Text></Text>
@@ -36,10 +36,13 @@ export default function ReportsScreen({ navigation }: RootTabScreenProps<'Report
                 <MonthlyExpenseGraph />
                 <MonthlyVsBudgeted />
             </ScrollView>
-        </Screen>
+        </View>
     );
 }
 
 const styles = StyleSheet.create({
-
+    screen: {
+        flex: 1,
+        backgroundColor: 'white',
+    },
 });
