+++ screens/UserAuthScreens/SignUpScreen.tsx
@@ -4,8 +4,8 @@ import { ActivityIndicator, Alert, ScrollView, StyleSheet } from 'react-native';
 import { FontAwesome } from '@expo/vector-icons';
 import { Text, View, RequiredField } from '../../components/Themed';
 import { RootStackScreenProps } from '../../types';
-import TextInput from '../../components/TextInput';
-import Button from '../../components/Button';
+import TextInput from '../../components/forms/TextInput';
+import Button from '../../components/buttons/Button';
 import Styles from "../../constants/Styles";
 import { styles, eyeIconSize } from './SignUpScreen.styles';
 import { CreateUserMutation, CreateUserDocument } from '../../components/generated';
