+++ screens/UserAuthScreens/WelcomeScreen.tsx
@@ -1,7 +1,7 @@
 import { GestureResponderEvent, StyleSheet, Image } from 'react-native';
 
 import { Text, View } from '../../components/Themed';
-import Button from '../../components/Button';
+import Button from '../../components/buttons/Button';
 import React from 'react';
 import { RootStackScreenProps } from '../../types';
 import { useAuth } from '../../hooks/useAuth';
